Major: 0.9.126
Bug Fixed:
1. Finally fixed the open dialog bug.

+++ src/plugin/knmusicplugin/plugin/knmusicbackendbass/knmusicbassanalysiser.cpp
@@ -44,7 +44,7 @@ bool KNMusicBassAnalysiser::analysis(KNMusicDetailInfo &detailInfo)
                                   BASS_MUSIC_DECODE);
 #endif
     //Check the channel.
-    if(NULL==analysisChannel)
+    if((DWORD)NULL==analysisChannel)
     {
         return false;
     }
